Handle super() calls.
super() calls in constructors are special, they have to be turned into
an initializer call. This only translates super() calls in immediate
constructor children to match Angular's existing Traceur-based
transpiler's behaviour. It still seems a bit risky, we might want to
print out a warning if we run into issues with it.

super.foo() style calls elsewhere do not need special treatment and are
translated as is.