Disable filtering out CONST, IMPLEMENTS, ABSTRACT annotations.
It turns out that dartanalyze will complain about unused imports if
these annotations are no longer present. Without real type information
it's not possible to detect those cases reliably, so it's easier for now
to just include the annotations in the output.

Long term, we might either want to ignore those warnings or use the type
information from TS on what's actually used or isn't.