Pass our TypeScript version to gulp-typescript.
This fixes the build as TSC started compiling about `let` statements in
ES5 targeting code. The newer `typescript.d.ts` itself contains `let`
statements, which may or may not be intentional, so we might have to
move to ES6 in the future. For the time being, this explicitly passes
in the TypeScript dependency to make sure our dependency versions are
half-way sane.