feat: defensive check that a symbol has some declaration.
Possibly due to incorrect TypeScript syntax, symbols might not have any
declarations at all. This adds some defensive code to report that as an error
condition.

The change is a bit larger than expected because we don't have a `ts.Node` down
in that function and need one to give a good location for the error.